Move Speaker model to correct package
