+++ app/src/main/java/net/squanchy/speaker/SpeakerService.java
@@ -5,6 +5,7 @@ import java.util.Comparator;
 import java.util.List;
 
 import net.squanchy.service.firebase.FirebaseSquanchyRepository;
+import net.squanchy.speaker.domain.view.Speaker;
 
 import io.reactivex.Observable;
