+++ app/src/main/java/net/squanchy/speaker/domain/view/Speaker.java
@@ -1,4 +1,4 @@
-package net.squanchy.speaker;
+package net.squanchy.speaker.domain.view;
 
 import com.google.auto.value.AutoValue;
